termutils: made environment query assume a default
If the terminal size cannot be determine by any of the 3 methods used in this
function, assume the default size for a terminal of (24 x 80).

The last block of code had an undeclared variable 'env', so if this block had
been executed it would not have worked.

Changed to using 'if cr is None' to conform with PEP8,

http://www.python.org/dev/peps/pep-0008/