termutils: added LinuxTerminalSize class
Added a class with functions for the different methods of querying the window
size. This made is easier to test each of the implementations.

I have moved the import statements out of the querying function; there is really
no siginificant benefit to lazy importing.

Introduced the contextlib package to cleanly open and close the controlling
terminal process in order to query its size.